+++ protected/humhub/modules/admin/views/authentication/authentication.php
@@ -24,82 +24,37 @@ $this->registerJsConfig('admin.space', [
 $userModule = Yii::$app->getModule('user');
 
 $this->beginContent('@admin/views/authentication/_authenticationLayout.php');
+?>
 
-echo Html::beginTag('div', ['class' => 'panel-body']);
+<div class="panel-body">
+    <?php $form = ActiveForm::begin(['id' => 'authentication-settings-form', 'acknowledge' => true]) ?>
 
-$form = ActiveForm::begin(['id' => 'authentication-settings-form', 'acknowledge' => true]);
+    <?= $form->errorSummary($model) ?>
 
-echo $form->errorSummary($model);
+    <?= $form->field($model, 'allowGuestAccess')->checkbox() ?>
+    <?= $form->field($model, 'internalAllowAnonymousRegistration')->checkbox() ?>
+    <?= $form->field($model, 'internalUsersCanInviteByEmail')->checkbox() ?>
+    <?= $form->field($model, 'internalUsersCanInviteByLink')->checkbox() ?>
+    <?= $form->field($model, 'internalRequireApprovalAfterRegistration')->checkbox() ?>
+    <?= $form->field($model, 'showRegistrationUserGroup')->checkbox() ?>
+    <?= $form->field($model, 'blockUsers')->checkbox() ?>
+    <?= $form->field($model, 'hideOnlineStatus')->checkbox() ?>
+    <?= $form->field($model, 'allowUserTopics')->checkbox(['data' => ['action-change' => 'admin.space.restrictTopicCreation']]) ?>
+    <?= $form->field($model, 'defaultUserIdleTimeoutSec')->textInput(['readonly' => $userModule->settings->isFixed('auth.defaultUserIdleTimeoutSec')]) ?>
+    <p class="help-block"><?= Yii::t('AdminModule.user', 'The default user idle timeout is used when user session is idle for a certain time. The user is automatically logged out after this time.') ?></p>
+    <?= $form->field($model, 'defaultUserProfileVisibility')->dropDownList(User::getVisibilityOptions(false), ['readonly' => (!Yii::$app->getModule('user')->settings->get('auth.allowGuestAccess'))]) ?>
+    <p class="help-block"><?= Yii::t('AdminModule.user', 'Only applicable when limited access for non-authenticated users is enabled. Only affects new users.') ?></p>
 
-echo $form->field($model, 'allowGuestAccess')->checkbox();
-echo $form->field($model, 'internalAllowAnonymousRegistration')->checkbox();
-echo $form->field($model, 'internalUsersCanInviteByEmail')->checkbox();
-echo $form->field($model, 'internalUsersCanInviteByLink')->checkbox();
-echo $form->field($model, 'internalRequireApprovalAfterRegistration')->checkbox();
-echo $form->field($model, 'showRegistrationUserGroup')->checkbox();
-echo $form->field($model, 'blockUsers')->checkbox();
-echo $form->field($model, 'hideOnlineStatus')->checkbox();
-echo $form->field($model, 'allowUserTopics')->checkbox(['data' => ['action-change' => 'admin.space.restrictTopicCreation']]);
-echo $form->field($model, 'defaultUserIdleTimeoutSec')
-    ->textInput(['readonly' => $userModule->settings->isFixed('auth.defaultUserIdleTimeoutSec')]);
+    <?php if (Yii::$app->getModule('user')->settings->get('auth.needApproval')) : ?>
+        <?= $form->field($model, 'registrationSendMessageMailContent')->widget(RichTextField::class, ['exclude' => ['oembed', 'upload']]) ?>
+        <?= $form->field($model, 'registrationApprovalMailContent')->widget(RichTextField::class, ['exclude' => ['oembed', 'upload']]) ?>
+        <?= $form->field($model, 'registrationDenialMailContent')->widget(RichTextField::class, ['exclude' => ['oembed', 'upload']]) ?>
+        <p class="help-block"><?= Yii::t('AdminModule.user', 'Do not change placeholders like {displayName} if you want them to be automatically filled by the system. To reset the email content fields with the system default, leave them empty.') ?></p>
+    <?php endif; ?>
 
-echo Html::tag(
-    'p',
-    Yii::t(
-        'AdminModule.user',
-        'The default user idle timeout is used when user session is idle for a certain time. The user is automatically logged out after this time.'
-    ),
-    ['class' => 'help-block']
-);
+    <hr/>
+    <?= Html::submitButton(Yii::t('AdminModule.user', 'Save'), ['class' => 'btn btn-primary', 'data-ui-loader' => '']) ?>
 
-echo $form->field($model, 'defaultUserProfileVisibility')
-    ->dropDownList(
-        User::getVisibilityOptions(false),
-        ['readonly' => (!Yii::$app->getModule('user')->settings->get('auth.allowGuestAccess'))]
-    );
-
-echo Html::tag(
-    'p',
-    Yii::t(
-        'AdminModule.user',
-        'Only applicable when limited access for non-authenticated users is enabled. Only affects new users.'
-    ),
-    ['class' => 'help-block']
-);
-
-if (Yii::$app->getModule('user')->settings->get('auth.needApproval')) {
-    echo $form->field($model, 'registrationSendMessageMailContent')->widget(
-        RichTextField::class,
-        ['exclude' => ['oembed', 'upload']]
-    );
-    echo $form->field($model, 'registrationApprovalMailContent')->widget(
-        RichTextField::class,
-        ['exclude' => ['oembed', 'upload']]
-    );
-    echo $form->field($model, 'registrationDenialMailContent')->widget(
-        RichTextField::class,
-        ['exclude' => ['oembed', 'upload']]
-    );
-
-    echo Html::tag(
-        'p',
-        Yii::t(
-            'AdminModule.user',
-            'Do not change placeholders like {displayName} if you want them to be automatically filled by the system. To reset the email content fields with the system default, leave them empty.',
-        ),
-        ['class' => 'help-block']
-    );
-}
-
-echo Html::tag('hr');
-
-echo Html::submitButton(
-    Yii::t('AdminModule.user', 'Save'),
-    ['class' => 'btn btn-primary', 'data-ui-loader' => ""]
-);
-
-ActiveForm::end();
-
-echo Html::endTag('div');
-
-$this->endContent();
+    <?php ActiveForm::end() ?>
+</div>
+<?php $this->endContent() ?>